refactor: change file structure of the project
