+++ src/Hooks/useGetHeroData.ts
@@ -1,6 +1,6 @@
 import { useRef, useState } from 'react';
 import SwapiService from 'services/SwapiService';
-import { PeopleRequestType } from 'view/Search/Search';
+import { PeopleRequestType } from 'components/Search/Search';
 
 export default function useGetHeroData(heroNumber: number) {
   const [heroData, setHeroData] = useState<PeopleRequestType>(Object);
