+++ src/components/AppRoutes.tsx
@@ -1,7 +1,7 @@
 // eslint-disable-next-line object-curly-newline
 import { Route, Routes, BrowserRouter, Navigate } from 'react-router-dom';
-import AppLayout from 'view/AppLayout/AppLayout';
-import MainPage from 'view/MainPage/MainPage';
+import AppLayout from 'components/AppLayout/AppLayout';
+import MainPage from 'components/MainPage/MainPage';
 
 function AppRoutes() {
   return (
