+++ src/components/Details/DetailsComponent.scss
