+++ src/components/Details/types.ts
@@ -1,0 +1,4 @@
+export type DetailsComponentPropsType = {
+  setShowDetails: React.Dispatch<React.SetStateAction<boolean>>;
+  heroNumber: number;
+};
