+++ src/components/ErrorBoundary/ErrorBoundary.scss
