+++ src/components/ErrorBoundary/types.ts
@@ -1,0 +1,7 @@
+export interface ErrorBoundaryState {
+  error: Error | null;
+}
+
+export interface ErrorBoundaryProps {
+  children: React.ReactNode;
+}
