+++ src/components/ErrorButton/types.ts
@@ -1,0 +1,3 @@
+export interface ErrorButtonState {
+  error: boolean;
+}
