+++ src/components/MainPage/MainPage.scss
