+++ src/components/MyLoader/MyLoader.scss
