+++ src/components/MyLoader/types.ts
@@ -1,0 +1,3 @@
+export type MyLoaderPropsType = {
+  stylesClassName: string;
+};
