+++ src/components/Search/Search.tsx
@@ -1,40 +1,7 @@
 import React from 'react';
 import './Search.scss';
-import SearchIcon from 'view/Search/SearchIcon';
-
-export type PeopleRequestType = {
-  name: string;
-  height: string;
-  mass: string;
-  hair_color: string;
-  skin_color: string;
-  eye_color: string;
-  birth_year: string;
-  gender: string;
-  homeworld: string;
-  films: string[];
-  species: [];
-  vehicles: string[];
-  starships: string[];
-  created: string;
-  edited: string;
-  url: string;
-};
-
-export type PeoplesRequestType = {
-  count: number;
-  next: string;
-  previous: null;
-  results: PeopleRequestType[];
-};
-
-export type SearchPropsType = {
-  searchFormHandler: (event: React.FormEvent<HTMLFormElement>) => void;
-  searchValue: string;
-  fetchError: string;
-  setFetchError: (value: string) => void;
-  setSearchValue: (value: string) => void;
-};
+import SearchIcon from 'components/Search/SearchIcon';
+import { SearchPropsType } from 'components/Search/types';
 
 export default function Search({
   setSearchValue,
