+++ src/components/SearchItems/NoDataComponent.scss
